added hook TooltipsActions
