+++ frontend/src/components/pages/MainComponent/MyTooltips/MyTooltips.jsx
@@ -8,15 +8,9 @@ import {
 import { Tooltip } from "antd";
 import { useTooltipsActions } from "./../../../../hooks/useTooltipsActions";
 
-export default function MyTooltips({
-    // needToDo,
-    task,
-    showModal,
-    deleteTask,
-    startTask,
-    completeTask,
-}) {
-    const { needToDo } = useTooltipsActions();
+export default function MyTooltips({ task, showModal }) {
+    const { needToDo, startTask, completeTask, deleteTask } =
+        useTooltipsActions();
 
     return [
         <Tooltip key="edit" title="Редактировать">
